# Improve performance of deny_* methods by not raising in prod environments