--- a/.changeset/gorgeous-spies-hang.md
+++ b/.changeset/gorgeous-spies-hang.md
@@ -1,0 +1,5 @@
+---
+'@primer/view-components': patch
+---
+
+Improve performance of the deny\_\* methods