--- a/demo/Gemfile
+++ b/demo/Gemfile
@@ -40,8 +40,8 @@ gem "lookbook", "~> 1.3.4" unless rails_version.to_f < 7
 
 group :development do
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-  gem "spring"
-  gem "spring-watcher-listen", "~> 2.0.0"
+  gem "spring", "~> 4.0"
+  gem "spring-watcher-listen", "~> 2.1"
   gem "hotwire-livereload", "~> 1.1"
   # Use JavaScript with ESM import maps [https://github.com/rails/importmap-rails]
   gem "importmap-rails"