--- a/test/test_helpers/ips_test_helper.rb
+++ b/test/test_helpers/ips_test_helper.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+# This code is exercised by rake bench
+# :nocov:
+
+require "benchmark/ips"
+
+module Primer
+  module IPSTestHelper
+    def measure_ips
+      suppress_output do
+        Benchmark.ips { |x| x.report { yield } }
+      end
+    end
+
+    def assert_more_ips(faster, slower, msg = nil)
+      assert faster.data.first[:ips] > slower.data.first[:ips], msg
+    end
+
+    private
+
+    def suppress_output
+      original_stderr = $stderr.clone
+      original_stdout = $stdout.clone
+      $stderr.reopen(File.new("/dev/null", "w"))
+      $stdout.reopen(File.new("/dev/null", "w"))
+      yield
+    ensure
+      $stdout.reopen(original_stdout)
+      $stderr.reopen(original_stderr)
+    end
+  end
+end