… demo/Gemfile.lock
… test/test_helpers/components/primer/deny_component.rb