@camertron 2022-11-22
@jonrohan that's a great idea. Maybe we could add a test that ensures the prod version is faster than the non-prod version. Let me see what I can do.